requirements.txt
selenium
webdriver_manager
python-dotenv
browser-cookie3

